pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
name = "firemerge"
description = "Firefly merge/import GUI"
version = "0.1.0"
readme = "README.md"
dependencies = [
    "selenium ~= 4.25.0",
    "aiohttp ~= 3.10.0",
    "python-dotenv ~= 1.0.0",
    "tomlkit ~= 0.13.0",
    "pydantic ~= 2.9.0",
]

[project.optional-dependencies]
dev = [
    "mypy==1.11",
]

[tool.setuptools]
packages = ["firemerge"]

[project.scripts]
firemerge-scrape = "firemerge.scrape:main"
firemerge = "firemerge.web:main"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]